Fix test access_list presentation
